230214 오전

+++ src/main/resources/static/js/board.js
@@ -4,7 +4,7 @@ let index = {
 			this.save();
 		});
 		$("#btn-delete").on("click",()=>{ // function(){}, ()=>{}를 사용한 이유는 this를 바인딩하기 위해서!!
-			this.deleteById();
+			this.deleteBoard();
 		});
 		$("#btn-update").on("click",()=>{ // function(){}, ()=>{}를 사용한 이유는 this를 바인딩하기 위해서!!
 			this.update();
@@ -44,19 +44,35 @@ let index = {
 		}); 
 	},
 	
-	deleteById:function(){
-		let id = $("#id").text();
-		
-		$.ajax({
-			type: "DELETE",
-			url: "/api/board/"+id,
-			dataType: "json" 
-		}).done(function(resp){
-			alert("삭제가 완료되었습니다.");
-			location.href = "/";
-		}).fail(function(error){
-			alert(JSON.stringify(error));
-		}); 
+	
+	deleteBoard:function deleteBoard(seq){
+			Swal.fire({
+			  title: '글을 삭제하시겠습니까???',
+			  text: "삭제하시면 다시 복구시킬 수 없습니다.",
+			  icon: 'warning',
+			  showCancelButton: true,
+			  confirmButtonColor: '#3085d6',
+			  cancelButtonColor: '#d33',
+			  confirmButtonText: '삭제',
+			  cancelButtonText: '취소'
+			}).then((result) => {
+			  if (result.value) {
+	              //"삭제" 버튼을 눌렀을 때 작업할 내용을 이곳에 넣어주면 된다. 
+					let id = $("#id").text();
+					
+					$.ajax({
+						type: "DELETE",
+						url: "/api/board/"+id,
+						dataType: "json" 
+					}).done(function(resp){
+						alert("삭제가 완료되었습니다.");
+						location.href = "/";
+					}).fail(function(error){
+						alert(JSON.stringify(error));
+					}); 
+	
+			  }
+			})
 	},
 	
 	update:function(){
